refactor PresetItem
