+++ console/src/components/RadioBase/PresetItem.less
@@ -1,0 +1,8 @@
+.slotItem {
+    display: flex;
+    justify-content: space-between;
+
+    .after {
+        padding: 18px;
+    }
+}
