+++ console/src/components/RadioBase/package.json
@@ -1,0 +1,3 @@
+{
+	"main": "RadioBase.js"
+}
